+++ backend/Back/models/DTOs/Request/RecompensaRequest.cs
@@ -4,7 +4,7 @@ public class RecompensaRequest
 {
     public string Nombre { get; set; } = string.Empty;
     public string? Descripcion { get; set; }
-    public decimal CostoPuntos { get; set; }
+    public double CostoPuntos { get; set; }
     public int Stock { get; set; }
     public bool Activa { get; set; } = true;
 }
